# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kodementor Api Documentation",
        "description": "Kodementor Api Documentation",
        "contact": {
            "name": "Vijay Rana",
            "email": "info@kodementor.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/chat-room": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Chat room",
                "description": "Chat room",
                "operationId": "ChatRoom",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "message": {
                                                        "type": "string"
                                                    },
                                                    "chat_room_id": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key_security_example": {
                "type": "http",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "scheme": "Bearer",
                "in": "header",
                "bearerFormat": "JWT"
            },
            "api_key_security_example_access": {
                "type": "http",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "scheme": "Bearer",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "oauth2_security_example": [
                "read",
                "write"
            ]
        }
    ]
}